chore: add shared component
