+++ apps/payments/e2e/specs/paymentRequests/paymentRequestStripe.spec.ts
@@ -130,7 +130,7 @@ test.describe("Payment Request with stripe provider", () => {
     const previewPayPage = new PreviewPayPage(newPage);
     await previewPayPage.checkHeader();
     await previewPayPage.checkAmount(mockAmount);
-    await previewPayPage.checkCustomAmountOptionNotVisible();
+    await previewPayPage.customAmountForm.checkCustomAmountOptionNotVisible();
     await previewPayPage.checkPaymentMethodHeader();
     await previewPayPage.checkPaymentMethodVisible("card");
     await previewPayPage.checkButton();
